+++ example-app/lib/services/mq/agenda.js
@@ -6,9 +6,9 @@ exports.open = () => {
   if (!agenda) {
     const { MONGO_URL, JOB_TYPES, JOB_MONGO_URL, JOB_COLLECTION } = global.CONFIG
 
-    const address = JOB_MONGO_URL === 'mongo' ? MONGO_URL : JOB_MONGO_URL || ''
+    const address = JOB_MONGO_URL || MONGO_URL
     const jobTypes = JOB_TYPES ? JOB_TYPES.split(',') : []
-    if (jobTypes.length && address) {
+    if (jobTypes.length && address && JOB_COLLECTION) {
       const Agenda = require('agenda')
       const connectionOpts = { // mongodb options
         db: {
@@ -21,12 +21,13 @@ exports.open = () => {
         }
       }
       agenda = new Agenda(connectionOpts)
-      
       jobTypes.forEach(type => {
         require(APP_PATH + '/jobs/' + type)(agenda)
       })
       agenda.start() // Returns a promise, which should be handled appropriately
       console.log('agenda started - job count=' + jobTypes.length)
+    } else {
+      console.log('agenda not started (JOB_MONGO_URL, JOB_TYPES, JOB_COLLECTION)', address, JOB_TYPES, JOB_COLLECTION)
     }
   }
   return this